Add open source implementation for SNS credentials
